server/requirements.txt
fastapi
databases[asyncpg]
python-jose[cryptography]
python-multipart
uvicorn[standard]